requirements.txt
langchain
streamlit
python-dotenv
transformers
langchain-community
torch
langchain-openai
sentence-transformers
langchain-groq